[ 2022. 02. 22 ] 1. Custom JPA Repository 생성 및 구현 2. 특정 날짜를 조건으로 하여 결제 내역을 가져오는 로직 구현
** 웹 서칭을 통하여 기능 구현은 하였으나 Spring-data-jpa 를 잘 활용한 것인지는 의문이 듬.

+++ src/main/java/com/coffee/vo/Order.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Table(name="payment_history")
 public class Order {
 	
-	public int cost;
+	public double cost;
 	@Id
 	public String orderid;
 	public String order_date;
